flask_python_aks_workload_identity_app/requirements.txt
Flask
azure-identity
azure-keyvault-secrets